example/app/requirements.txt
South
django-extensions
